add permissions for project and middleware for project's routes

+++ src/routes/api.php
@@ -35,11 +35,11 @@ Route::group(['namespace' => 'API'], function () {
 			Route::delete('users/{user}', 'ProjectController@destroy')->middleware(['permission:users-destroy']);
 
              // Projects
-           Route::post('projects', 'ProjectController@store');
-           Route::get('projects', 'ProjectController@index');
-           Route::get('projects/{id}', 'ProjectController@show');
-           Route::put('projects/{project}', 'ProjectController@update');
-           Route::delete('projects/{project}', 'ProjectController@destroy');
+           Route::post('projects', 'ProjectController@store')->middleware(['permission:projects-index']);
+           Route::get('projects', 'ProjectController@index')->middleware(['permission:projects-show']);
+           Route::get('projects/{id}', 'ProjectController@show')->middleware(['permission:projects-store']);
+           Route::put('projects/{project}', 'ProjectController@update')->middleware(['permission:projects-update']);
+           Route::delete('projects/{project}', 'ProjectController@destroy')->middleware(['permission:projects-destroy']);
 
 
 		});
